% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nfl_Weights.R
\docType{data}
\name{Nfl_Weights}
\alias{Nfl_Weights}
\title{NFL Weights Data}
\format{
A data frame with 52 rows and 1 variable:
 \describe{
     \item{Weights}{the weights of the football players
      measured in pounds.}
      }
}
\source{
{Chicago Tribune}
}
\usage{
data(Nfl_Weights)
}
\description{
Contains the weights of 52 randomly selected NFL
football players. This data set appears in Chapter 8,
section 8.5 problem 13.
}
\examples{
data(Nfl_Weights)
mean(Nfl_Weights$Weights)

data(Nfl_Weights)
tint(x = Nfl_Weights$Weights, CL=0.9)
}
\keyword{datasets}
