% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_zint_function.R
\name{p.zint}
\alias{p.zint}
\title{One-Sample z-Interval for Proportions}
\usage{
p.zint(X, n, CL = 0.95)
}
\arguments{
\item{X}{the numerator input for the proportion}

\item{n}{the denominator input for the proportion}

\item{CL}{the confidence level of the interval, defaults to 0.95}
}
\value{
The function returns the values of the sample proportion,
    the sample size, the standard error, the critical value, and
    the lower and upper bounds for the confidence interval.
}
\description{
This function performs a simple 1-prop two-sided z-interval.
}
\examples{
p.zint(400,10000,CL=0.98)

p.zint(X=632,n=800)

}
