% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tint_function.R
\name{tint}
\alias{tint}
\title{One-Sample t-Interval}
\usage{
tint(x, s, xbar = NULL, n = NULL, CL = 0.95)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data. Not needed if user has
access to the sample statistics.}

\item{s}{the sample standard deviation.}

\item{xbar}{the sample mean.}

\item{n}{the sample size.}

\item{CL}{the confidence level of the interval, defaults to 0.95. Can
be changed to any value between 0 and 1.}
}
\value{
The function returns the values of the sample mean,
    the sample size, the sample standard deviation, the critical value,
    the degrees of freedom, and the lower and upper bounds for the
    confidence interval.
}
\description{
This function performs a simple one-sample two-sided t-interval.
}
\examples{
Sal=c(0.593, 0.142, 0.329, 0.691, 0.231, 0.793, 0.519, 0.392,
    0.418, 0.409, 0.587, 0.446, 0.379, 0.542, 0.138, 0.662,
    0.525,0.354, 0.496, 0.725)
tint(Sal,CL=.99)

tint(xbar=116.9,s=21.7,n=10,CL=0.95)

}
